Extract test about DateUtils.format localization to its own class
